requirements.txt
beautifulsoup4==4.9.1
colorama==0.4.3
cryptography==3.4.7
discord.py==2.3.2
dnspython==1.16.0
pymodm==0.4.3
python-dateutil==2.8.1
python-dotenv==0.20.0
requests==2.21.0
